Extract DebugLogger to shared library
