+++ AnSAM/Services/GameListService.cs
@@ -6,7 +6,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using System.Xml;
 using System.Diagnostics;
-using AnSAM.Services;
+using CommonUtilities;
 
 namespace AnSAM.Services
 {
