+++ AnSAM/Services/IconCache.cs
@@ -1,11 +1,11 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using AnSAM.Services;
 using System.IO;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using CommonUtilities;
 
 namespace AnSAM.Services
 {
