+++ AnSAM/Steam/GameImageUrlResolver.cs
@@ -1,6 +1,6 @@
 using System;
 using System.IO;
-using AnSAM.Services;
+using CommonUtilities;
 
 namespace AnSAM.Steam
 {
