+++ MyOwnGames/MainWindow.xaml.cs
@@ -28,6 +28,7 @@ using WinRT.Interop;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
+using CommonUtilities;
 
 namespace MyOwnGames
 {
