+++ MyOwnGames/Services/FailedDownloadService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml.Linq;
+using CommonUtilities;
 
 namespace MyOwnGames.Services
 {
