+++ MyOwnGames/SteamApiService.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using System.Globalization;
 using MyOwnGames.Services;
+using CommonUtilities;
 
 namespace MyOwnGames
 {
