+++ RunGame/Services/AchievementIconService.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
 // Service no longer constructs BitmapImages directly
+using CommonUtilities;
 
 namespace RunGame.Services
 {
