+++ RunGame/Steam/SteamGameClient.cs
@@ -8,7 +8,7 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading;
 using Microsoft.Win32;
-using RunGame.Services;
+using CommonUtilities;
 
 namespace RunGame.Steam
 {
